Replay & File Event Handler stuff bs (Not working yet)
Tweaked the replay so that whenever the game ends. A prompt shows to enter initials, once enters, asks the users if they want to replay or go to menu.

Code may be a bit messy right now since I'm tryna figure out how to create a text document and keep the scores in there.